pulled check_path into iiqtools.utils.generic so other scripts can use it
